calendar terminado

+++ faemood/src/components/pages/Profile.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { auth } from "../../config/firebase";
-import { onAuthStateChanged, signOut } from "firebase/auth";
+import { onAuthStateChanged, signOut, reauthenticateWithCredential, EmailAuthProvider, EmailAuthCredential } from "firebase/auth";
 import '../../styles/Profile.css';
 import Modal from "../elements/Modal";
 import ModalDelete from "../elements/ModalDelete";
@@ -11,6 +11,7 @@ function Profile() {
     const [loading, setLoading] = useState(true); // estado de carga
     const [isModalVisible, setIsModalVisible] = useState(false); // estado del modal de logout
     const [isDeleteAccModalVisible, setisDeleteAccModalVisible] = useState(false)
+    const [password, setPassword] = useState("")
 
   
 
@@ -36,15 +37,39 @@ function Profile() {
         setIsModalVisible(false); // oculta el modal
     };
 
-    // función para borrar cuenta
+    const reauthenticateUser = async () => {
+        const user = auth.currentUser
+        const credential = EmailAuthProvider.credential(user.email, password)
 
-    const deleteUserAcc = async () => {
+        try {
+            await reauthenticateWithCredential(user, credential)
+            console.log("Reautenticación exitosa");
+            return true
+        } catch (err) {
+            console.error("Error en la reautenticación:", err)
+            return false
+        }
+    };
+
+
+
+    // función para borrar cuenta
+    const deleteUserAcc = async () => { // para que el usuario pueda borrar su cuenta debe de haberse registrado recientemente (funcion reuthenticateUser)
         const user = auth.currentUser
         if (user) {
             try {
-                await user.delete()
+                const reuthenticated = await reauthenticateUser()
+                if(reuthenticated) {
+                    try {
+                        await user.delete()
+                        console.log("Acc deleted")
+                    } catch (err) {
+                        console.error("Error al eliminar la cuenta: ", err)
+                    }
+                } else {
+                    console.log("No se puede eliminar la cuenta")
+                }
 
-                console.log("Account deleted")
             } catch (err) {
                 console.error("Error:", err)
             }
@@ -90,10 +115,7 @@ function Profile() {
     return (
         <div className="profilContainer">
             <h2 className="titP">User Profile</h2>
-            <p>image</p>
-            <p><strong>Username:</strong></p>
             <p><strong>Email:</strong> {user.email}</p>
-            <p><strong>Email:</strong> {user.username}</p>
             <button className="logout" onClick={handleLogout}>Log out</button>
             <Modal
                 isVisible={isModalVisible}
